custom-agent/pyproject.toml
[project]
name = "custom-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.1",
    "strands-agents>=1.12.0",
    "uvicorn[standard]>=0.37.0",
]
